Address concerns
KAFKA-268
